backend/pyproject.toml
[project]
name = "boinchub"
dynamic = ["version"]
description = "A BOINC account manager intended for personal use."
readme = "README.md"
license = "MIT"
authors = [{ name = "Jason Lynch", email = "jason@aexoden.com" }]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.1",
    "argon2-cffi>=23.1.0",
    "fastapi[standard]>=0.115.12",
    "lxml>=5.4.0",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "pydantic-xml>=2.16.0",
    "python-jose[cryptography]>=3.4.0",
    "sqlmodel>=0.0.24",
    "types-lxml>=2025.3.30",
]

[project.scripts]
boinchub = "boinchub:main"
boinchub-admin = "boinchub.cli.admin:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "boinchub/__about__.py"

[dependency-groups]
dev = ["ruff>=0.11.6"]

[tool.ruff]
exclude = ["alembic"]
line-length = 120

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = ["COM812", "D203", "D213", "ISC001"]

[tool.ruff.lint.flake8-copyright]
notice-rgx = "(?i)SPDX-FileCopyrightText: \\d{4}"

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel", "pydantic_xml.BaseXmlModel"]

[tool.ruff.lint.isort]
lines-between-types = 1

[tool.ruff.lint.pydocstyle]
convention = "pep257"
